Now we're getting somewhere!
The "House Houdini" bug is squashed an insect!
Also, the game state for the plants is being moved to the server. On
placement of a house, a plant is added underneath it. In addition, the
plants expand outwardly on the server and the game state is sent to the
players. The only problem is that the plants don't have proper growth,
and they seem to blink in and out with each update from the server.
Investigating this further now...