Game Killer Bug Squashed with Enthusiasm
There was a bug where I passed values to 'expandPlant' in the order y, x
when it should have been x, y. This caused the wrong array index to be
used while plants were expanding and depending on where your plants were
the game server would crash. Now it doesn't. This also fixed the
blinking in and out of the plants. Still though, the plant growth is
weirdly diagonal. Looking further into this...