Plants on server, need better growth algorithm
The plant growth was actually success fully moved over on the last
commit. We just didn't realize the flaw in our plant growth algorithm
because of the initial state of plants. The plants will inherently
growth to the top left much faster. We need to come up with a better
algorithm for this growth so that it is more equally distributed. This
might also be fixed inherently when we add tilling mechanics.