Debugging Report
This commit is mostly for explanation purposes in case I do not solve
the problem tonight. I have narrowed it down to the drawGrid function.
It seems that allPlayers for the given plant pid (playerid) is
undefined. I can't see how though. I have some debugging console logs
which seem to confirm it. But then if I access the console on the
browser and print out the members of allPlayers, the matching key is in
there.