Powerups baseline laid
Laid a baseline for plant-based powerups to be implemented. Waterbucket
attempted to implement as area of effect based upgrade for 10 seconds.
Sometimes it causes an exception to be thrown when the water bucket is
used though. I think it is because some of the plants in the plants
array are not yet defined for some reason.