Fixed some stability issues
Fixed crashes due to trying to access 'deadColor' of undefined objects.
Also added a statement to request the users list before the player sends
a request for the rest of the game state. This is necessary because
without it, the grid cannot be drawn correctly at the name entering
screen.